refactor(api): move certification-point-of-contact from iam to deprecated
