+++ api/src/deprecated/infrastructure/serializers/jsonapi/certification-point-of-contact.serializer.js
