+++ api/tests/deprecated/unit/infrastructure/serializers/jsonapi/certification-point-of-contact.serializer.test.js
@@ -1,8 +1,8 @@
-import { certificationPointOfContactSerializer } from '../../../../../../src/identity-access-management/infrastructure/serializers/jsonapi/certification-point-of-contact.serializer.js';
+import { certificationPointOfContactSerializer } from '../../../../../../src/deprecated/infrastructure/serializers/jsonapi/certification-point-of-contact.serializer.js';
 import { expect } from '../../../../../test-helper.js';
 import { domainBuilder } from '../../../../../tooling/domain-builder/domain-builder.js';
 
-describe('Unit | Identity Access Management | Serializer | JSONAPI | certification-point-of-contact-serializer', function () {
+describe('Unit | Deprecated | Serializer | JSONAPI | certification-point-of-contact-serializer', function () {
   describe('#serialize()', function () {
     it('converts a CertificationPointOfContact model into JSON API data', function () {
       // given
